+++ src/main/scala/xitrum/handler/inbound/UriParser.scala
@@ -35,7 +35,10 @@ class UriParser extends SimpleChannelInboundHandler[HandlerEnv] {
       case NonFatal(e) =>
         val msg = "Could not parse query params URI: " + request.getUri
         Log.warn(msg, e)
-        ctx.channel.close()
+        // https://github.com/xitrum-framework/xitrum/issues/508#issuecomment-72808997
+        // Do not close channel without responding status code.
+        // Nginx decides that upstream is down if upstream drop connection without responding status code.
+        BadClientSilencer.respond400(ctx.channel, "Server could not parse params in URI")
     }
   }