+++ CHANGELOG.rst
@@ -1,5 +1,7 @@
 3.22:
 
+* `#503 <https://github.com/xitrum-framework/xitrum/issues/503>`_
+  Make respondSockJsXxx return Future like other respondXxx methods
 * `#459 <https://github.com/xitrum-framework/xitrum/issues/459>`_
   Optimize route collecting: Ignore more packages that obviously don't contain routes
 * `#495 <https://github.com/xitrum-framework/xitrum/issues/495>`_